refactor(registry): the command-to-distiller decision lives in one place (#194)
`distillers::route` was twenty `if matches!(base, …)` blocks, which is the
standing violation of the SSOT guardrail in `CLAUDE.md`. #323 closed the
half with a demonstrated cost. What kept the rest open is that a
`&str -> distiller` map cannot express it: `cargo` routes by subcommand,
`git` and `gh` decline on a flag, and the JVM and Dart arms split on
whether the command line mentions a test.

So the registry answers with a `Distillation` enum that has a
`Passthrough` variant, and `route` is one `match`. The ten
command-reading predicates moved with it, so `passes_through_verbatim`,
`wraps_another_command`, `lists_containers`, `lists_kubectl_names`,
`git_enumerates_files`, `is_vcs_list_command` and `cargo_subcommand` sit
beside the mapping that uses them. The two hooks call
`registry::passes_through_verbatim` now.

Order is preserved exactly, because order is the behaviour: `mvn` and
`gradle` appear in both the JVM and build-tool arms, `kubectl … exec` has
to reach `wraps_another_command` before the cloud arm claims it, and the
grep arm has to answer before `passes_through_verbatim`.

Verified by differential, not by reading. 5,227 commands, being the 5,125
distinct ones recorded in `~/.omni/omni.db` plus 102 hand-picked for arms
the corpus does not cover, each run through eleven probes shaped for
eleven distiller families, before and after. All 57,497 outputs are
byte-identical. The harness was then shown to have teeth by routing
`psql` to the generic distiller, which moved 16 commands.

Two things were deleted rather than moved. The npm arm's subcommand check
returned `JsTsDistiller` on both branches, so it decided nothing. And the
routing arm carried its own copy of the enumeration rationale that
`passes_through_verbatim`'s doc already states, one line of which had
gone stale on #326. Test count is unchanged at 614; four moved to sit
with the functions they cover.

Closes #194
