+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **The benchmark behind the published headline still counted terminal bytes, so the README overstated savings by 36 points (#324)**: `tests/bench_replay.rs` read `SELECT command, raw_input FROM execution_traces` with no agent filter. On the reporting installation that corpus is 68% `terminal` by bytes, 888 traces carrying 86.0 MB of TTY output no model ever receives, against 13.9 MB from `claude_code` and 25.7 MB from `aider`. #212 established exactly this and fixed it in `omni stats`, which now prints "Terminal output is excluded" on its own line; the harness that produces the number the README quotes never got the same fix. Replaying the same pipeline over the same database, one run each: **everything, 126.0 MB to 26.3 MB, 79.1% net**; **model-facing only, 40.1 MB to 22.7 MB, 43.3% net**. The harness replays the model-facing population by default now, prints which population it used, and `OMNI_BENCH_ALL=1` asks for the wider one. Both are worth having, because the gap between them is the mistake, and hiding one of the two is how it comes back.
 
 ### Changed
+- **The command-to-distiller decision moved into `pipeline/registry.rs`, where `CLAUDE.md` says it belongs (#194)**: `distillers::route` was twenty `if matches!(base, …)` blocks, the standing violation of the SSOT guardrail, and #323 had already closed the half of it with a demonstrated cost. What kept the rest open was that a `&str -> distiller` map cannot express it: `cargo` routes by subcommand, `git` and `gh` decline on a flag, and the JVM and Dart arms split on whether the command line mentions a test. So the registry answers with a `Distillation` enum that has a `Passthrough` variant, and `route` is one `match`. The ten command-reading predicates went with it, so `passes_through_verbatim`, `wraps_another_command`, `lists_containers`, `lists_kubectl_names`, `git_enumerates_files`, `is_vcs_list_command` and `cargo_subcommand` now live beside the mapping that uses them; the two hooks call `registry::passes_through_verbatim` instead. Order is preserved exactly, because order is the behaviour here: `mvn` and `gradle` appear in both the JVM and build-tool arms, `kubectl … exec` has to reach `wraps_another_command` before the cloud arm claims it, and the grep arm has to answer before `passes_through_verbatim`. Verified by differential rather than by reading: every one of **5,227 commands** (5,125 distinct ones recorded in `~/.omni/omni.db`, plus 102 hand-picked for the arms the corpus does not cover, from `cargo tree` to `az aks command invoke` to `git show --stat`) was run through eleven probes shaped for eleven distiller families, before and after, and the 57,497 outputs are byte-identical. The harness was then shown to have teeth by routing `psql` to the generic distiller, which moved 16 commands. One duplicated rationale block was deleted rather than moved: the routing arm carried its own copy of the enumeration reasoning that `passes_through_verbatim`'s doc already states, and one line of that copy had already gone stale on #326. Test count is unchanged at 614, with four moved to sit beside the functions they cover.
 - **README cut from 370 lines to 219, and `docs/PERFOMANCE.md` deleted**: ten `##` sections meant half the page was never read while still standing between a visitor and the install command. `How it works`, `Architecture` and the Memory OS section moved to `docs/ARCHITECTURE.md`; the corpus, per-command table, fixtures and latency tables moved to `docs/BENCHMARKS.md`, which also absorbs the method block `tests/bench_replay.rs` cites. `Why This Matters` and `Integrations` were deleted outright, because the hero already says which agents are supported and the intro already says what the problem is. `docs/PERFOMANCE.md` went with them: it published 58.9% over an 1,810-trace corpus that #324 superseded, and a second performance document with a different corpus is how a stale number survives a correction. Nothing measured was dropped, and every figure that moved to `docs/` is still reachable in one click from Benchmarks. All six translations follow the same shape.
 - **New section: how to read a savings number, including ours**: five questions with our answers, which is the only comparison worth making. A feature-by-feature table against named competitors was drafted and thrown away after checking their READMEs: reversibility is not a differentiator (headroom caches originals for retrieval, lean-ctx documents five recovery paths), passthrough is not either (lean-ctx ships 270 passthrough rules), and rtk already publishes its own caveat that cutting 90% of bash output is not cutting a bill by 90%. Three of the five rows would have been false. What no tool in the category publishes is the share of calls where it saved nothing, so that is the row we lead with, and it names nobody.
 - **README leads with what OMNI does well instead of with its worst aggregate**: not one figure changed, and none was rounded. The old hero opened on a category description and buried the per-command results under a 43.3% whole-mix average that includes thousands of commands OMNI deliberately does nothing to. The new one opens on `git` 89%, `cargo` 91%, `kubectl` 77%, which is where the saving actually is, and the 43.3% still appears in Benchmarks with the same corpus and the same caveat. "90.0% of calls saved nothing" now reads "OMNI acts on 1 command in 10, and adds zero bytes to the other 9", which is the same fact stated as the design decision it is. "Latency is a real cost, not zero" leads with the measured 21 ms and keeps the 205 MB row underneath it. Two guarantees were promoted out of a table into prose, because RewindStore recovery and the `Option<String>` return that makes fabrication unrepresentable are the two things no competing compressor offers, and they were the smallest text on the page. The line about publishing the share of calls where we did nothing stays, sharpened: a tool claiming 90% on every command is telling you it summarised something you needed. All six translations follow, and three stale figures went with them: the `git diff` 45% and `docker build` 37% that predate the retrieval marker, a `kubectl get pods` row still promising a `35 pods | 30 running, 5 error` summary the pipeline stopped producing, and Arabic's 82/308/276 ms latency block.
