+++ src/distillers/mod.rs
@@ -34,421 +34,6 @@ pub trait Distiller: Send + Sync {
     ) -> Option<String>;
 }
 
-/// The subcommand of a `cargo` invocation, skipping env-var prefixes, the path
-/// the binary was called by, a `+toolchain` override, and any leading flags.
-///
-/// `None` when there is no subcommand at all (`cargo`, `cargo --version`), which
-/// routes to passthrough — the honest answer for output we cannot classify.
-fn cargo_subcommand(command: &str) -> Option<&str> {
-    let mut tokens = command.split_whitespace().skip_while(|t| {
-        // `RUST_BACKTRACE=1 cargo test`, and the cargo path itself.
-        let is_env_assignment = t.contains('=') && !t.starts_with('-');
-        let is_cargo = std::path::Path::new(t.trim_matches(['"', '\'', '`']))
-            .file_name()
-            .and_then(|n| n.to_str())
-            .map(|n| n == "cargo" || n == "cargo.exe")
-            .unwrap_or(false);
-        is_env_assignment || !is_cargo
-    });
-    tokens.next()?; // the cargo token itself
-
-    // `cargo +nightly tree`, `cargo --offline tree`
-    tokens.find(|t| !t.starts_with('-') && !t.starts_with('+'))
-}
-
-fn extract_base_executable(command: &str) -> String {
-    let tokens = shell_split_tokens(command, 8);
-    if tokens.is_empty() {
-        return String::new();
-    }
-
-    let mut i = 0usize;
-    while i < tokens.len() {
-        let t = tokens[i].as_str();
-        if t == "env" || t == "command" {
-            i += 1;
-            continue;
-        }
-        if looks_like_env_assignment(t) {
-            i += 1;
-            continue;
-        }
-        return tokens[i].clone();
-    }
-
-    String::new()
-}
-
-/// Commands whose output `distill_with_command` hands back byte-for-byte, and
-/// which the hooks must therefore also exempt from the collapse fallback.
-///
-/// The four halves reach the same rule for different reasons:
-///
-/// * **Enumeration** (`ls`, `find`, `ps`, `wc`, `df`, `du`, `stat`, `tree`, bare
-///   `env`) lists distinct data one datum per line. The items *are* the answer
-///   and there is no noise to drop (#198/#200).
-/// * **Bare interpreters** (`python`, `python3`, `ruby`) run an arbitrary
-///   program whose stdout is the answer, not a build log (#190).
-/// * **File readers** (`cat`, `head`, `tail`, `sed`, `awk`) emit whatever the
-///   file holds, so any shape a summariser recognises in it is a coincidence
-///   (#235/#236).
-/// * **The caller's own filter** (`grep`, `rg`, `ag`) already selected these
-///   lines by pattern, so a second selection cannot know what the first was
-///   after (#316/#326).
-///
-/// The single list is the point. A passthrough returns its input, so it can
-/// never beat `beats_guardrail`, so the hooks treat it as a distiller that
-/// punted and collapse it anyway — which put 40 distinct data rows behind one
-/// `[N similar lines collapsed]` marker and reported 95.7% saved (#214). Adding
-/// a passthrough arm without adding it here is that bug, so both live in one
-/// predicate rather than in a routing arm and a guard that can drift apart.
-///
-/// The filter half is the one exception to "hands back byte-for-byte", and it is
-/// deliberate. `route` reaches the grep arm **before** this predicate, so the
-/// grep distiller still runs: it hoists repeated paths losslessly when that
-/// shrinks the payload, and returns the input when it does not. Membership here
-/// is what stops the hooks collapsing the payload on the second outcome, and it
-/// is load-bearing rather than defensive: with these three names removed, 121
-/// matched lines came back as `120 INFO entries (collapsed from 120 lines)`.
-/// Measured over 214 recorded `grep`/`rg` traces before adding it: the hoist
-/// fires on 46 of them for 13.4 KB, which is why the arm stays, and 28 more came
-/// back carrying a `[Partial signal]` marker over lines the pattern had matched.
-pub fn passes_through_verbatim(command: &str) -> bool {
-    let base_exec = extract_base_executable(command);
-    let base = std::path::Path::new(&base_exec)
-        .file_name()
-        .and_then(|n| n.to_str())
-        .unwrap_or(base_exec.as_str());
-    matches!(
-        base,
-        "ls" | "tree"
-            | "find"
-            | "ps"
-            | "df"
-            | "du"
-            | "stat"
-            | "wc"
-            | "python"
-            | "python3"
-            | "ruby"
-            // `make` is a composite runner, not a build tool. A target runs
-            // whatever its recipe says, which is usually several programs, and
-            // their output arrives concatenated with no delimiter — `make check`
-            // is tsc, then eslint, then pytest. `BuildDistiller` owned the whole
-            // buffer and could only speak for one of them: on a run where each
-            // gate found something, 431 B came back as 111 B headed
-            // `Build: 1 errors, 0 warnings`, with the eslint finding deleted and
-            // a headline that counted neither the warning nor the failed test
-            // (#129). That is the same rule already applied to `python`, `ruby`
-            // and `az aks command invoke`: a thing that runs an arbitrary
-            // program has no grammar of its own, so the honest answer is the
-            // input. `gcc`, `clang`, `cmake` and `cargo` still route to the
-            // distiller when invoked directly, so a compile still compresses;
-            // only the wrapper passes through. Measured before choosing: 0 of
-            // 9,832 recorded commands start with `make `, so this trades a
-            // compression that has never fired here for a false claim that does.
-            | "make"
-            // File readers. Their stdout is whatever the file holds, so there is
-            // no grammar to distil and every shape a summariser recognises in it
-            // is a coincidence. `cat docs/DEVELOPMENT.md` — 127 lines of prose —
-            // came back as `tree: 127 entries` because one code block inside the
-            // document drew a directory layout (#236), and `sed -n '285,340p'` of
-            // a Swift file lost 37 of 56 lines to the enumeration cut (#235).
-            // Same reasoning as the interpreters above: an arbitrary payload.
-            | "cat"
-            | "head"
-            | "tail"
-            | "sed"
-            | "awk"
-            // The caller's own filter. The pattern already picked these lines, so
-            // the collapse fallback is a second filter that cannot know what the
-            // first was looking for (#316/#326). Measured with these three names
-            // taken back out: 121 matched lines came back as `120 INFO entries
-            // (collapsed from 120 lines)` over a `119 lines omitted` marker. The
-            // grep distiller still runs, because `route` reaches it above the
-            // call to this predicate, so a lossless hoist is unaffected; this
-            // only governs what happens when it punts.
-            | "grep"
-            | "rg"
-            | "ag"
-            // Format renderers. Their whole job is to emit something a later step
-            // parses, which is the format-safe contract stated in `AGENTS.md`,
-            // and the collapse fallback would leave that payload unparseable.
-            // `kubectl get pod -o json | jq -r '...'` lost three of four lines
-            // before this (#269).
-    )
-        // The reshaping tails `registry::owning_tail` routes to, read from the
-        // one list rather than repeated here (#277, #194). Naming a stage as the
-        // payload's owner only helps if that stage is then handled, and none of
-        // these has a grammar: `cut` and `column` project or lay out columns,
-        // `tr` and `base64` rewrite bytes, `xargs` runs some other program.
-        // Keeping the two in step by comment is what let `gh api … | base64 -d`
-        // reach the generic distiller and get cut, on the first run of #277.
-        || crate::pipeline::registry::RESHAPING_TAILS.contains(&base)
-        // `extract_base_executable` strips a leading `env`/`command` wrapper, so
-        // bare `env` and `command env` both leave base empty — match on any `env`
-        // token in that case rather than only the first word (review of #203).
-        || (base.is_empty() && command.split_whitespace().any(|t| t == "env"))
-        // A git command asked for a file list by flag rather than by name (#231).
-        // It belongs here and not only in the routing arm: a passthrough that the
-        // collapse fallback then folds up is #214, and the file rows of a
-        // `--stat` all share a shape.
-        || (base == "git" && git_enumerates_files(command))
-        // A container listing, for the same reason (#233).
-        || lists_containers(command)
-        // A kubectl listing that has no columns to summarise (#301).
-        || lists_kubectl_names(command)
-        // A wrapper whose stdout belongs to whatever it ran inside (#234).
-        || wraps_another_command(command)
-}
-
-/// Commands whose stdout is some other program's, not their own.
-///
-/// `az aks command invoke` does not produce `az` output: it runs an arbitrary
-/// shell command inside a cluster and hands back that program's stdout. Routing
-/// on the first executable sent it to a cloud distiller with no grammar for
-/// whatever ran, and a four-section run came back as nine rows of the first
-/// section plus `[Partial signal]`, with the TLS probe that was the whole point
-/// of the call discarded. It cost a 40 second round trip against a private
-/// cluster to find that out (#234). `kubectl exec` has form here too: #112 was
-/// its output summarised as `docker logs: 9 lines, no errors detected`.
-///
-/// Same rule as `python`, `python3` and `ruby` in #190: a thing that runs an
-/// arbitrary program has no grammar of its own, so the honest answer is the
-/// input.
-///
-/// The alternative was parsing the inner command and routing on that, which
-/// keeps the distillation on `ssh host 'cargo build'`. Measured before choosing:
-/// 59 of 8,032 recorded commands are wrappers, 0.7%, so that parser and its
-/// failure modes would be bought for less than one call in a hundred. If the
-/// share grows, revisit it with the same query.
-fn wraps_another_command(command: &str) -> bool {
-    let mut tokens = command.split_whitespace().map(|t| t.trim_matches('"'));
-    let base = tokens
-        .next()
-        .and_then(|w| std::path::Path::new(w).file_name()?.to_str());
-
-    match base {
-        // `ssh host '<cmd>'`, but not `ssh host` alone, which is interactive and
-        // produces nothing to distil either way.
-        Some("ssh") => true,
-        // `kubectl exec … -- <cmd>`, `docker exec … <cmd>`, `podman exec …`.
-        Some("kubectl") | Some("docker") | Some("podman") => {
-            command.split_whitespace().any(|t| t == "exec")
-        }
-        // `az aks command invoke -c '<cmd>'`.
-        Some("az") => command.contains("command invoke"),
-        _ => false,
-    }
-}
-
-/// `docker ps`, `podman ps` and `docker container ls` list containers one per
-/// row, and every column on that row is the answer: name, image, port mapping,
-/// uptime.
-///
-/// The summariser kept the wrong half. `docker: 5 containers | 3 running, 2
-/// exited` names only the exited ones by construction, so a fixture with three
-/// running containers lost every name, image and port it had, at a reported 91.3%
-/// saving and with no marker, because nothing thought anything was lost (#233).
-/// An agent runs `docker ps` to find a running container's port far more often
-/// than to find a dead one.
-///
-/// Same reasoning that put `ls`, `find`, `ps`, `df`, `du`, `stat` and `wc` on the
-/// list above, and `ps` is the closest relative: a busy machine's process table
-/// dwarfs any container list and has been verbatim since 0.6.6. The `-a` case on
-/// a host with hundreds of dead containers is the one with real noise, and it
-/// stays bounded by `MAX_OUTPUT_BYTES`, which cuts with a marker that says what
-/// it removed.
-///
-/// Deliberately narrow. `docker build` and `docker logs` are logs with real
-/// noise and keep their summarisers, and `docker compose ps` is a different
-/// command whose output shape has not been measured here.
-fn lists_containers(command: &str) -> bool {
-    let mut tokens = command.split_whitespace().map(|t| t.trim_matches('"'));
-    let base = tokens
-        .next()
-        .and_then(|w| std::path::Path::new(w).file_name()?.to_str());
-    if !matches!(base, Some("docker") | Some("podman")) {
-        return false;
-    }
-    match tokens.next() {
-        Some("ps") => true,
-        Some("container") => matches!(tokens.next(), Some("ls") | Some("ps")),
-        _ => false,
-    }
-}
-
-/// `kubectl` asked for a list of names rather than a table of state.
-///
-/// `kubectl config get-contexts -o name` prints one identifier per row and
-/// nothing else, so there is no status column to summarise and no noise to drop:
-/// the identifiers *are* the answer, the same rule as `ls`, `find`, `docker ps`
-/// and every other enumeration on this list. `distill_kubectl_generic` keeps only
-/// rows that match its critical or `configured|created|unchanged|deleted`
-/// vocabulary, and a context name matches neither, so a 20 row list came back as
-/// the first 10 with the rest dropped. Because the estate sorts alphabetically,
-/// the cut landed on the tail and every `k8s-*` production cluster was the half
-/// that went, which is worse than a random ten: the delivered list looks complete
-/// and contains no production context at all (#301).
-///
-/// Narrow on purpose. `kubectl get pods` keeps its distiller, because a pod table
-/// has a `STATUS` column with real noise in it and a fingerprint (`READY` +
-/// `RESTARTS`) that says so. This covers the forms whose output has no columns:
-/// `-o name` on any subcommand, and the three subcommands that only ever
-/// enumerate.
-fn lists_kubectl_names(command: &str) -> bool {
-    let mut tokens = command.split_whitespace().map(|t| t.trim_matches('"'));
-    let base = tokens
-        .next()
-        .and_then(|w| std::path::Path::new(w).file_name()?.to_str());
-    if base != Some("kubectl") {
-        return false;
-    }
-    let rest: Vec<&str> = tokens.collect();
-
-    // `-o name`, `--output name`, `-oname`, `--output=name`.
-    let asks_for_names = rest
-        .windows(2)
-        .any(|w| matches!(w[0], "-o" | "--output") && w[1] == "name")
-        || rest.iter().any(|t| *t == "-oname" || *t == "--output=name");
-
-    asks_for_names
-        || matches!(
-            rest.first().copied(),
-            Some("api-resources") | Some("api-versions")
-        )
-        || (rest.first().copied() == Some("config")
-            && matches!(
-                rest.get(1).copied(),
-                Some("get-contexts") | Some("get-clusters") | Some("get-users")
-            ))
-}
-
-fn looks_like_env_assignment(token: &str) -> bool {
-    let Some((key, _value)) = token.split_once('=') else {
-        return false;
-    };
-    if key.is_empty() {
-        return false;
-    }
-    key.chars().all(|c| c.is_ascii_alphanumeric() || c == '_')
-}
-
-fn shell_split_tokens(input: &str, max_tokens: usize) -> Vec<String> {
-    #[derive(Clone, Copy)]
-    enum Mode {
-        None,
-        Single,
-        Double,
-        Backtick,
-    }
-
-    let mut tokens: Vec<String> = Vec::new();
-    let mut current = String::new();
-    let mut mode = Mode::None;
-    let mut chars = input.chars().peekable();
-
-    while let Some(ch) = chars.next() {
-        if tokens.len() >= max_tokens {
-            break;
-        }
-
-        match mode {
-            Mode::None => match ch {
-                '\'' => mode = Mode::Single,
-                '"' => mode = Mode::Double,
-                '`' => mode = Mode::Backtick,
-                '\\' => {
-                    if let Some(next) = chars.next() {
-                        current.push(next);
-                    }
-                }
-                c if c.is_whitespace() => {
-                    if !current.is_empty() {
-                        tokens.push(std::mem::take(&mut current));
-                    }
-                    while matches!(chars.peek(), Some(p) if p.is_whitespace()) {
-                        chars.next();
-                    }
-                }
-                _ => current.push(ch),
-            },
-            Mode::Single => {
-                if ch == '\'' {
-                    mode = Mode::None;
-                } else {
-                    current.push(ch);
-                }
-            }
-            Mode::Double => match ch {
-                '"' => mode = Mode::None,
-                '\\' => {
-                    if let Some(next) = chars.next() {
-                        current.push(next);
-                    }
-                }
-                _ => current.push(ch),
-            },
-            Mode::Backtick => match ch {
-                '`' => mode = Mode::None,
-                '\\' => {
-                    if let Some(next) = chars.next() {
-                        current.push(next);
-                    }
-                }
-                _ => current.push(ch),
-            },
-        }
-    }
-
-    if !current.is_empty() && tokens.len() < max_tokens {
-        tokens.push(current);
-    }
-
-    tokens
-}
-
-/// `gh` and `glab` are wrappers: `gh pr list` prints the enumeration
-/// `VcsDistiller` was written to summarise, and `gh api …` prints whatever the
-/// endpoint returned. The distiller cuts to the first 10 lines on size alone,
-/// with no grammar check, so a Swift file fetched through
-/// `gh api … | base64 -d | sed -n '285,340p'` lost 37 of its 56 lines under
-/// `... [37 more items — use --limit to see more]` — a flag neither `gh api`
-/// nor `sed` has, so the suggested recovery could not be run (#235). The same
-/// cut took a Homebrew cask down to 10 lines and stripped its blank lines on
-/// the way (#226).
-///
-/// `--limit` is the tell: it belongs to the `list` subcommands and nothing else,
-/// so those are the only outputs this distiller can honestly claim.
-fn is_vcs_list_command(command: &str) -> bool {
-    command.split_whitespace().any(|t| t == "list")
-}
-
-/// `--stat`, `--numstat`, `--name-only` and `--name-status` exist only to make
-/// git emit a file list. The flag is a reliable signal in a way the subcommand
-/// is not: `git show` without one prints a diff and is legitimately distillable,
-/// and with one it prints the enumeration that *is* the answer.
-///
-/// `git show --stat` reached `distill_log`, because the input holds no
-/// `diff --git` and no `On branch`. There the `--oneline` subject matched
-/// `RE_GIT_LOG_HASH` and was kept, while the stat rows fell past every arm to
-/// the rule that drops body lines before the next commit (#199) — correct for
-/// `git log`, wrong for `git show`, where that position holds the payload. The
-/// fail-open guard below `distill_log` then did not fire, because `result` was
-/// non-empty: one line *had* been recognised. That is #228 again in a different
-/// distiller — a partial recognition disarming a guard whose condition is
-/// "recognised nothing" — and it published a 79.9% saving for a commit whose
-/// four files, three of them screenshots, had all vanished from the answer
-/// (#231).
-fn git_enumerates_files(command: &str) -> bool {
-    command.split_whitespace().any(|t| {
-        matches!(
-            t,
-            "--stat" | "--numstat" | "--shortstat" | "--name-only" | "--name-status"
-        )
-    })
-}
-
 /// The same zero-state rule as `Distiller::distill`'s `None`, for the helpers
 /// several layers below it that still hand a `String` back to their caller.
 ///
@@ -476,12 +61,17 @@ pub fn distill_with_command(
     route(segments, input, command, session).unwrap_or_else(|| input.to_string())
 }
 
+/// One `match` over the registry's decision. Every arm of what used to be here
+/// is now a variant of `registry::Distillation`, which is where `CLAUDE.md` says
+/// the command-to-behaviour mapping belongs (#194).
 fn route(
     segments: &[crate::pipeline::OutputSegment],
     input: &str,
     command: &str,
     session: Option<&crate::pipeline::SessionState>,
 ) -> Option<String> {
+    use crate::pipeline::registry::Distillation;
+
     // A chain's stdout belongs to several programs and arrives as one stream, so
     // there is no honest way to hand it to the distiller named by the first of
     // them: `git status && echo === && find .` came back as the git one-liner
@@ -489,240 +79,32 @@ fn route(
     // several pass through.
     let command = crate::pipeline::registry::sole_output_command(command)?;
 
-    // 1. Resolve pipeline profile (though we match command here too)
-    let _profile = crate::pipeline::registry::resolve_profile(command);
-
-    // Phase 1: Try exact command prefix match
-    let base_exec = extract_base_executable(command);
-    let base = std::path::Path::new(&base_exec)
-        .file_name()
-        .and_then(|n| n.to_str())
-        .unwrap_or(base_exec.as_str())
-        .to_string();
-    let cmd_lower = command.to_lowercase();
-
-    // Git subcommand routing
-    if base == "git" {
-        if git_enumerates_files(command) {
-            return None;
+    match crate::pipeline::registry::resolve_distiller(command) {
+        Distillation::Passthrough => None,
+        Distillation::Git => git::GitDistiller.distill(segments, input, session),
+        Distillation::Database => database::DatabaseDistiller.distill(segments, input, session),
+        Distillation::Security => security::SecurityDistiller.distill(segments, input, session),
+        Distillation::Vcs => vcs::VcsDistiller.distill(segments, input, session),
+        Distillation::Test => test::TestDistiller.distill(segments, input, session),
+        Distillation::Build => build::BuildDistiller.distill(segments, input, session),
+        Distillation::JsTs => jsts::JsTsDistiller.distill(segments, input, session),
+        Distillation::Cloud(tool) => {
+            cloud::CloudDistiller { tool: &tool }.distill(segments, input, session)
         }
-        return git::GitDistiller.distill(segments, input, session);
+        // The only distiller allowed near a payload the caller's pattern already
+        // selected: it hoists repeated paths and hands the input back when that
+        // does not shrink it, so neither outcome can lose a match (#316/#326).
+        Distillation::UserFiltered => Some(system_ops::distill_user_filtered(input)),
+        Distillation::SystemOps => system_ops::SystemOpsDistiller.distill(segments, input, session),
+        Distillation::Generic => generic::GenericDistiller.distill(segments, input, session),
     }
-
-    // Database tools
-    if matches!(
-        base.as_str(),
-        "psql" | "mysql" | "sqlite3" | "pg_dump" | "redis-cli"
-    ) {
-        return database::DatabaseDistiller.distill(segments, input, session);
-    }
-
-    // Security scanners
-    if matches!(
-        base.as_str(),
-        "semgrep" | "trivy" | "snyk" | "hadolint" | "gosec" | "bandit"
-    ) {
-        return security::SecurityDistiller.distill(segments, input, session);
-    }
-
-    // GitHub/VCS CLIs
-    if matches!(base.as_str(), "gh" | "hub" | "glab") {
-        if !is_vcs_list_command(command) {
-            return None;
-        }
-        return vcs::VcsDistiller.distill(segments, input, session);
-    }
-
-    // Java/JVM — use BuildDistiller (sudah ada)
-    if matches!(
-        base.as_str(),
-        "java" | "javac" | "mvn" | "mvnw" | "gradle" | "gradlew"
-    ) {
-        if cmd_lower.contains("test") {
-            return test::TestDistiller.distill(segments, input, session);
-        }
-        return build::BuildDistiller.distill(segments, input, session);
-    }
-
-    // Flutter/Dart
-    if matches!(base.as_str(), "flutter" | "dart") {
-        if cmd_lower.contains("test") || cmd_lower.contains("analyze") {
-            return test::TestDistiller.distill(segments, input, session);
-        }
-        return build::BuildDistiller.distill(segments, input, session);
-    }
-
-    // Build tools → BuildDistiller
-    if matches!(
-        base.as_str(),
-        "cargo"
-            | "cmake"
-            | "gcc"
-            | "g++"
-            | "clang"
-            | "rustc"
-            | "go"
-            | "pip"
-            | "pip3"
-            | "ruff"
-            | "mypy"
-            | "black"
-            | "rake"
-            | "rubocop"
-            | "dotnet"
-            | "gradle"
-            | "mvn"
-            | "pytest"
-            | "rspec"
-            | "phpunit"
-    ) {
-        // `cargo` is routed by subcommand, not by executable (#170). Most of
-        // cargo is not a build: `cargo tree` prints a dependency tree that *is*
-        // the answer, and `BuildDistiller` summarised 21.4 KB of it as
-        // `Build: ok` — 9 bytes — reported as a 100% reduction. A command whose
-        // output is data must not be handed to a distiller that only knows how
-        // to count compile steps.
-        if base == "cargo" {
-            return match cargo_subcommand(command) {
-                Some("test" | "bench") => test::TestDistiller.distill(segments, input, session),
-                Some(
-                    "build" | "check" | "run" | "clippy" | "rustc" | "fix" | "doc" | "install",
-                ) => build::BuildDistiller.distill(segments, input, session),
-                // tree, metadata, search, pkgid, locate-project, vendor, add,
-                // remove, update, publish, … — output is the answer, hand it back.
-                _ => None,
-            };
-        }
-
-        // Tapi test → TestDistiller
-        if cmd_lower.contains("test")
-            || cmd_lower.contains("pytest")
-            || matches!(base.as_str(), "pytest" | "rspec" | "phpunit")
-        {
-            return test::TestDistiller.distill(segments, input, session);
-        }
-        return build::BuildDistiller.distill(segments, input, session);
-    }
-
-    // JS/TS ecosystem → JsTsDistiller
-    if matches!(
-        base.as_str(),
-        "vitest" | "playwright" | "tsc" | "eslint" | "prettier" | "jest" | "esbuild" | "vite"
-    ) {
-        return jsts::JsTsDistiller.distill(segments, input, session);
-    }
-    // npm/pnpm/yarn/bun: check subcommand
-    if matches!(base.as_str(), "npm" | "npx" | "pnpm" | "yarn" | "bun") {
-        if cmd_lower.contains("test")
-            || cmd_lower.contains("vitest")
-            || cmd_lower.contains("jest")
-            || cmd_lower.contains("playwright")
-        {
-            return jsts::JsTsDistiller.distill(segments, input, session);
-        }
-        // install/build → still JsTs ecosystem (pnpm install, npm run build)
-        return jsts::JsTsDistiller.distill(segments, input, session);
-    }
-
-    // The caller's own filter, which is in `passes_through_verbatim` and so has
-    // to be answered before it. `SystemOpsDistiller` cannot serve this: it
-    // dispatches on the *shape* of the payload, and a grep result has whatever
-    // shape the file had. When none of its detectors matched it fell to
-    // `distill_fallback`, which scores by noise and cut 8 of 15 lines a pattern
-    // had explicitly matched, under a `[Partial signal]` marker (#316/#326).
-    // The grep path is the only one that cannot do that: it hoists repeated
-    // paths and hands the input back whenever hoisting does not shrink it.
-    if matches!(base.as_str(), "grep" | "rg" | "ag") {
-        return Some(system_ops::distill_user_filtered(input));
-    }
-
-    if passes_through_verbatim(command) {
-        return None;
-    }
-
-    // Cloud & infra → CloudDistiller
-    if matches!(
-        base.as_str(),
-        "docker"
-            | "podman"
-            | "kubectl"
-            | "helm"
-            | "terraform"
-            | "tofu"
-            | "aws"
-            | "gcloud"
-            | "az"
-            | "doctl"
-    ) {
-        return cloud::CloudDistiller { tool: &base }.distill(segments, input, session);
-    }
-
-    // Enumeration commands list distinct data, one datum per line, with no noise
-    // to drop — the paths, filenames, processes and mounts *are* the answer. A
-    // command audit (docs/COMMAND_AUDIT.md) measured every one of these either
-    // dropping items (`find` 68/98 paths #198, `ls` every filename, `ps` most
-    // processes, `wc` per-file counts) or *growing* the output (`env` +50%,
-    // `df` +16%). There is nothing to distill losslessly here, so hand the output
-    // back verbatim rather than truncate it to a shorter, plausible, incomplete
-    // list (#200). `grep`/`rg` are NOT here: their distiller hoists the repeated
-    // path losslessly and keeps every match.
-    //
-    // The bare interpreters ride the same predicate: `python3 -c "..."` handed to
-    // BuildDistiller fabricated `Build: ok` for any script that printed no
-    // error line (#190). They sit here rather than in an arm of their own so the
-    // hooks' collapse exemption cannot drift away from the routing (#214). No
-    // `contains("test")` shortcut to TestDistiller for them, either: it matched
-    // inside a `-c` code argument or a path segment (`ruby /proj/contest/x.rb`),
-    // and TestDistiller fabricates too, which is #190 wearing another distiller's
-    // name. Real runners are handled upstream, where `signals/tools/pytest.toml`
-    // and `mypy.toml` are TOML-first and shadow this path for `python -m pytest`.
-    // System ops → SystemOpsDistiller
-    if matches!(
-        base.as_str(),
-        "curl"
-            | "wget"
-            | "sort"
-            | "uniq"
-            | "tar"
-            | "zip"
-            | "unzip"
-            | "apt"
-            | "apt-get"
-            | "brew"
-            | "yum"
-            | "dnf"
-    ) {
-        return system_ops::SystemOpsDistiller.distill(segments, input, session);
-    }
-
-    // Phase 2: Fallback to generic distiller
-    generic::GenericDistiller.distill(segments, input, session)
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::pipeline::registry::passes_through_verbatim;
     use crate::pipeline::scorer;
-
-    #[test]
-    fn reads_the_cargo_subcommand_past_prefixes_and_flags() {
-        assert_eq!(cargo_subcommand("cargo tree"), Some("tree"));
-        assert_eq!(cargo_subcommand("cargo tree --depth 1"), Some("tree"));
-        assert_eq!(
-            cargo_subcommand("RUST_BACKTRACE=1 cargo test"),
-            Some("test")
-        );
-        assert_eq!(cargo_subcommand("cargo +nightly build"), Some("build"));
-        assert_eq!(cargo_subcommand("cargo --offline tree"), Some("tree"));
-        assert_eq!(
-            cargo_subcommand("/usr/local/bin/cargo clippy"),
-            Some("clippy")
-        );
-        // No subcommand at all — routes to passthrough, which is the honest
-        // answer for output we cannot classify.
-        assert_eq!(cargo_subcommand("cargo"), None);
-        assert_eq!(cargo_subcommand("cargo --version"), None);
-    }
 
     /// From #170: `cargo tree` prints a dependency tree that *is* the answer.
     /// Routing every `cargo` command to `BuildDistiller` summarised 21.4 KB of
@@ -874,28 +256,6 @@ mod tests {
     }
 
     #[test]
-    fn test_extract_base_executable_handles_quotes_and_env_prefixes() {
-        assert_eq!(extract_base_executable("git diff"), "git");
-        assert_eq!(
-            extract_base_executable("\"/usr/local/bin/cargo\" build"),
-            "/usr/local/bin/cargo"
-        );
-        assert_eq!(extract_base_executable("'cargo' test"), "cargo");
-        assert_eq!(
-            extract_base_executable("`/usr/bin/python3` -V"),
-            "/usr/bin/python3"
-        );
-        assert_eq!(
-            extract_base_executable("RUST_BACKTRACE=1 cargo test"),
-            "cargo"
-        );
-        assert_eq!(
-            extract_base_executable("env FOO=1 \"/path/to/git\" status"),
-            "/path/to/git"
-        );
-    }
-
-    #[test]
     fn vitest_zero_parse_passes_through_instead_of_claiming_success() {
         // #143 / #115: a Vite dev server (no tests) must never distill to a
         // green `vitest: ✓ 0/0 passed`. With nothing parsed, fail open.
@@ -1024,34 +384,6 @@ mod tests {
                 "{cmd} produces the inner program's stdout, which OMNI has no grammar for"
             );
         }
-    }
-
-    /// The wrapper rule must not swallow the tools themselves.
-    #[test]
-    fn claims_only_the_wrapping_subcommands() {
-        assert!(wraps_another_command("kubectl exec pod -- ls"));
-        assert!(wraps_another_command("az aks command invoke -c 'ls'"));
-        assert!(wraps_another_command("ssh host 'ls'"));
-
-        assert!(!wraps_another_command("kubectl get pods"));
-        assert!(!wraps_another_command("az aks show -n cluster"));
-        assert!(!wraps_another_command("docker build ."));
-        assert!(!wraps_another_command("cargo test"));
-    }
-
-    /// The counter-case, so this is not "never distil docker". A build log and a
-    /// container log both have real noise and keep their summarisers.
-    #[test]
-    fn claims_only_the_container_listing_subcommands() {
-        assert!(lists_containers("docker ps"));
-        assert!(lists_containers("/usr/local/bin/docker ps -a"));
-        assert!(lists_containers("podman container ls"));
-
-        assert!(!lists_containers("docker build ."));
-        assert!(!lists_containers("docker logs app"));
-        assert!(!lists_containers("docker compose ps"));
-        assert!(!lists_containers("kubectl get pods"));
-        assert!(!lists_containers("ps aux"));
     }
 
     #[test]
